// src/app/models/request-changeaddress.model.ts
export interface IRequestChangeAddress {
    AddressType?: string;
    Address1?: string;
    Address2?: string;
    City?: string;
    ZipCode?: string;
    Country?: string;
    State?: string;
    Division?: string;
    District?: string;
    Thana?: string;
    Phone?: string;
    Mobile?: string;
    Mobile2?: string;
    Mobile3?: string;
    Mobile4?: string;
    Mobile5?: string;
    Email?: string;
    Fax?: string;
    Telex?: string;
    Swift?: string;
    Web?: string;
    ContactPersonNm?: string;
    ContactPersonDesig?: string;
    ContactPersonPhone?: string;
    ContactPersonMobile?: string;
    ContactPersonEmail?: string;
    UpozilaCitycorp?: string;
    UnionMuni?: string;
    VillWard?: string;
    IsCitycorp?: boolean;
}