// src/app/models/user_address.model.ts
export interface IUserFullAddress {
    customerID?:            string;
    addressTypeId?:         string;
    addressType?:           string;
    address1?:              string;
    address2?:              string;
    city?:                  string;
    zipCode?:               string;
    countryId?:             string;
    stateLGID?:             string;
    districtLGID?:          string;
    thanaLGID?:             string;
    countryName?:           string;
    divisionName?:          string;
    districtName?:          string;
    thanaName?:             string;
    phone?:                 string;
    mobile?:                string;
    mobile2?:               string;
    mobile3?:               string;
    mobile4?:               string;
    mobile5?:               string;
    email?:                 string;
    fax?:                   string;
    telex?:                 string;
    swift?:                 string;
    web?:                   string;
    contact_person_nm?:     string;
    contact_person_desig?:  string;
    contact_person_phone?:  string;
    contact_person_mobile?: string;
    contact_person_email?:  string;
    upozila_citycorp_id?:   null;
    upozila_citycorp_nm?:   null;
    union_muni_id?:         null;
    union_muni_nm?:         null;
    vill_ward_id?:          null;
    vill_ward_nm?:          null;
    ps_id?:                 null;
    ps_nm?:                 null;
    is_citycorp?:           null;
}